agregar panel libro y comic

+++ src/main/java/mx/anahuac/programacion_multimedia/Comic.java
@@ -3,7 +3,7 @@ package mx.anahuac.programacion_multimedia;
 public class Comic extends Libro{
     private String tipo_coloreado;
 
-    public Comic(int id, String name, String autor, String editorial, int precio, int inventario, String tipo_coloreado) {
+    public Comic(int id, String name, String autor, String editorial, Float precio, int inventario, String tipo_coloreado) {
         super(id, name, autor, editorial, precio, inventario);
         this.tipo_coloreado = tipo_coloreado;
     }
